Credentials Secret Storage (#2)
This addresses a potential security concern and shows how to integrate a secret storage in a x-platform way into the Xojo Post Build Scripts.

Instead of having a codesigning credential stored in a configuration file it will be retrieved from the secret storage and passed to the Docker Container via Environment Variable.

The Xojo Example Project includes an example which supports:
- macOS: Keychain
- Linux: Gnome Keyring
- Windows: Windows Credential Manager